Refactor about chat modal
